requirements.txt
kafka-python
datetime